[FEAT/FIX] Replace: Explicitly allow userinput
If replacements are defined before it shouldn't ask the user for any
more except when explicitly allowing userinput with '-manualinput'.

This fixes a problem that when -fix was provided it couldn't be
automated because it still asked if manual inputs are requested. This
changes the default behavior a bit, because previously manual input was
enabled when replacements via the command line were defined. But it
doesn't make sense that -manualinput is only required when using -fix.
This actually enables the bot to run automatically without user
interference using command line replacements and still allows the user
to specify replacements via the normal input.

Also removed explicit concatenation of two static strings and updated
documentation that -fix can now be used together with other parameters
like -nocase or additional replacements.

Change-Id: I199ed9e070b59424cb7251cc98ffd0a7ceff5b3f