[IMPROV] Add additional documentation
This adds a few missing docstrings, so that some files are now not
failing the non-voting jenkins test.

Repurposed the tests so that the normal flake8 test is now also doing
the docstring tests because the docstrings-mandatory test was doing the
normal flake8 tests anyway. The docstrings-mandatory test now contains
all files which already pass the non-voting docstrings test.

Change-Id: I213f1d05235a7f242a0776e44d37f8f89a8437c9