Fix solve_disambiguation
Currently solve_disambigation breaks frequently on English Wikipedia
pages.  It is instantiating a Link object with links obtained from
a regex match of the page text, and assuming the Link object doesnt
raise exceptions after instantiation.

Force a Link.parse() on the link text to force errors to occur at the
beginning of processing so the error can be caught and link skipped.

Bug: T85057
Change-Id: I2c874080620a5c777f5bd192421d3fcaa920f219