Update DeletionBot for undelete, implement tests
Update DeletionBot to support i18n messages for undelete, pass
mark=True to Page.delete() if -always is set (gets rid of prompt)

Add tests for Site.undelete and the deletionbot script.

Fix issues with Page.undelete and Site.undelete_page when an error
is raised if None is passed for revisions/timestamps.

Fix unhandled exception with delete.py if undeleting a page that
exists,

Bug: T74733
Bug: T57255
Bug: T62572
Change-Id: Ie474bc289951492f7fabc5ea53f63732df7acfdf