Add Generator Factory Tests
Added tests under TestFacoryGenerator for Allpages, Prefixing, and
RegexFilter PageGenerators. Also added a test for passing
only `-ns` to a GeneratorFactory.

The `default` for each for each tests the limit
and basic functionallity (Prefixing tests whether
the page starts with the prefix and RegexFilter
tests whether the returned pages match the regex).

The `ns` tests test whether the pages returned are in
the specified namespace.

`test_regexfilter_ns` is expected to fail due to bug T85389.
Essentially all results from `-titleregex` will have a namespace
of 0 and will return an empty generator if called with a namespace
other than 0.

Change-Id: I8a482bac50d983bb57d70468ce5bc61b2e031a81