[FIX] MediaWikiVersion: Support non-wmf versions
Previously it didn't require to match the complete version number. If
now a version '1.25alpha' was used in MediaWikiVersion it just ignored
the 'alpha' part. This now requires that the complete version number is
matched and does allow several different version number suffixes:

  * 'wmf<number>'
  * 'alpha'
  * 'beta<number>'
  * 'rc<number>'/'-rc.<number>'/'-rc<number>'/'rc.<number>'

It follows in that order, so an alpha version is considered newer (and
is thus using newer API) than a wmf version, but older than any beta
version. All suffixes are considered older than a version number without
a suffix. If the suffix doesn't follow in this scheme the version number
is invalid.

Because alpha releases might not represent the newest wmf version the
family can now define a 'force_version' number which must return a valid
version number for MediaWikiVersion. If it's not empty that version
number will be used at all times.

This in theory breaks the API as MediaWikiVersion has no 'wmf_version'
attribute, but that shouldn't be used anyway outside of that class. It
has been replaced by '_dev_version' which maps those above to a
'subversion' (e.g. wmf42 = (0, 42)). The stable releases (without a
suffix) use (4,) for simplification of the comparision.

As family's version method shouldn't be used directly anymore it's
marked as deprecated.

Change-Id: Id8b37d16765bae5007456dc9923dc5c81daca4e5