[IMPROV] Use commons for claim-type commonsMedia
The commonsMedia claim type only works for Wikimedia Commons so allowing
in theory any image repository doesn't make sense. To make it more
dynamic it's using a dictionary to exchange the conversion of the claim
value depending on the claim type. Without that it wouldn't be possible
to do the wikibase tests in dry mode because it would try to create a
Site which isn't allowed usually.

Change-Id: I727edf1120b7dc75cea72e679e16f122ac4de3ad