Add orderless page title assert to fix broken test
Split assertPagelistTitles into assertPageTitlesEqual and
assertPageTitlesCountEqual, with semantics similar to assertEqual
and assertCountEqual, however the generator is only used until
it has more items than the expected number of titles.

Create new test module 'tests' to test features of the test system.
It is disabled by default.

Fix pagegenerator tests which misused assertPagelistTitles 'titles' arg.

Change the reflinks xml tests to use assertPageTitlesCountEqual,
as the use of preloading means the order can vary.

Improve the xmlreader asserts to be more precise, to assert that
there are no order problems in xmlreader, and fix a unicode bug
in an expectedFailure test.

Change-Id: I00ceb32ac468b98f418d0fbe7ab1f0748e15b352