[FEAT] WikidataBot: Allow direct usage on Wikidata
The current WikidataBot requires that the generator works on pages and
that the ItemPages are created from them. But if the generator is
iterating over ItemPages from Wikidata it should use those directly.

There are three possible modes: It either always uses the
ItemPage.fromPage method or ItemPage constructor or it uses
ItemPage.fromPage only when the page is not in the item namespace.

Change-Id: I8b3f27c9e559dfbeb4a557a0b209b2e978ada5aa