Skip search test for sites using Wikia Search
The search API module is disabled on Wikia sites.
The method they use to disable the search module
is not yet known.

Skip this test on any site that is using extension
Wikia Search, so that this test is still triggered on
any other site which has disabled the search module.

Bug: T77968
Change-Id: I1e55fbc8ec77fdaea51b9565078a28622c4a260a