[FIX] Data: Use site's language and not code
Previously date.py was using the site's code which is not always the
same as the language, especially when they are projects which don't have
extra sites for the different languages. This is now using the site's
language.

To get the site's language a siteinfo request must be done so the two
tests which were failing previously must be now non-dry tests.

Bug: T85645
Change-Id: I8b7efd0cdd2ef5fa65ab57c2f47415932cb7d34b