Fetch patrol token for MW < 1.20
Get patrol token for versions < MW 1.20.

Since patrol was added in v1.14, until v1.16, the patrol token is the
same as the edit token.

For MW 1.17-19, the patrol token must be obtained from the query list
recentchanges.

Tests for patrol token in the different MW versions added.

Incompatibility of site version with requested version added.

Bug: T85727
Change-Id: Ibbba1f8e0093f28d8cefd86a71f225938bc4a8a1