[FIX] Pagegen test: Allow same chars anywhere
The test_regexfilter_default test requires that a page begins with the
same character at least twice. This is obviously not possible on
wikidata because all main namespace pages begin with Q and a digit which
are never the same. Thus the test never ends because it can't find 10
pages. This removes the requirement that the first two chracters need to
be the same, although this could still cause freezes if a wiki is very
small and doesn't contain 10 pages which have at least the same
character twice.

Those tests aren't executed by travis with wikidata configured so it
doesn't happen there currently.

Change-Id: I312334bc202894c7eb224c5f46ec32b725c2c961