[FIX] Editor: Support spaces, delete always and warn on \n (etc)
This support spaces in the editor file name and warns if it contains a
string literal like \n with Windows because the backslash was used as a
path delimiter without escaping them.

The pywikibot.editor.TextEditor.command method has been deprecated,
because outside scripts don't require access anyway and this makes it
possible to return a list of command segments.

This now deletes the temporary file on any case, also when there were no
changes done. And it uses the more secure 'mkstemp' function.

Bug: T86481
Change-Id: If7e9feee49acd1147c5554cdd74a7ad0758397c5