[FIX] gui: Fix import errors of gui module
The gui imports are invalid and thus it wasn't possible to use the
tkinter editor. The relative import in editor.py was neither in Python 3
nor in Python 2 valid and introduced with
728524b6a94c56a9fe92f338bfa5a8149e0a6ab5. The import in
terminal_interface_base.py is valid in Python 2 but not in Python 3 so
changed that into an absolute also.

Change-Id: I6fb408452ff7878e67b3edb953a02526af97c0e3