Prevent hang in LiveRCPageGenerator
The socketio client is being disconnected, but continues to poll
for new items indefinitely.

Detect disconnection and stop polling, making LiveRCPageGenerator
fail to fetch the requested total number of items instead of
stalling.

Correct item counter so that the thread quits after total
instead of total+1, reducing the chance of a disconnect.

These changes allows other tests to run after the disconnection,
lowering the impact and frequency of this bug.

Bug: T85717
Change-Id: I0748938b5fbe22a70e032f909c61d3c5bd2b7034