[FEAT] bot: add open_webbrowser
This adds a open_webbrowser method to the bot module to open the
webbrowser with the given page. The Page object now also supports to get
the full URL and the Family object can now returns a URI now with the
correct hostname and path prefix selected depending on the fact that SSL
is used.

Change-Id: Ic4f340b42981121216af3ab1f6cee00289fb6b47