[FEAT] Request: Support booleans
This supports the values to be booleans so that of an argument is set to
True, that it is added with an empty string and arguments set to False.
This only applies to a bool in an list or tuple of the length 1.

It changes how the values are internally stored: Instead of being casted
into string on defining the value (which removes the information whether
it was a bool or not) all values are stored how they were defined
(although maybe converted into a list if they didn't support iteration)
and are converted into a string only when the http URI is built.

It also adds a 'OptionSet' which is a list of booleans and creates then
a list of pipe-separated entries.

Change-Id: I1a61095168544334a0876aff927c4aa26e5fa125