[FIX] Flake8 on Python 3 fixes
This also changes how match_images in the match_images scripts return
the match value. Instead of being between 100 and 0 it's not between 1.0
and 0.0 so that the complete script is now using that range. It's in
theory breaking but because the name was pep8ified when converted to
core a few days ago it was broken anyway (and the old non-pep8 function
could be added to support the old 0-100 range if necessary). But because
of that it does now only say two images are the same, when they are
equal to 80 % and not 0.8 %.

This also renames the job from flake83 to flake8-py3.

Change-Id: I95ed0de30d2993d41a70c9b47b76c6461f40e8a0