[FIX] ui tests: Monkey patch pwb's UI class
This monkey patches the ui class used in pywikibot.bot. It is necessary
as on loading the pywikibot package it is creating a logger and handlers
which use the current std* streams (which are at that moment sys.std*)
as the target streams. This is intercepting a print to the console and
changing the targetStream (which was defined on creation of the handler)
to a internal buffer stream. Also reading from the console is
intercepted and redirected to returning the value of the internal
stream.

This is only patching stderr and stdout so if any other targetStream is
used it'll crash (intentionally so we can determine what is happening
there).

The ui_tests are now also activated so they are run by Travis.

Change-Id: I388f6253a401ef3be49f00b945e5414e47638109