Python 2.6 collections backports & failover
collections.Counter and .OrderedDict are a Python 2.7+ feature.
On Python 2.6, use backport package `future` by default, but also
use ordereddict and Counter packages if they are installed.

When a suitable implementation doesn't exist, provide a fake
class which raises an exception on use, so that it isn't
failing on import.

Bug: T72707
Change-Id: I265ee4726d124b8e5dfa8ec2e2f9314b2c2132e7