[FIX] reflinks: Use BytesIO for binary data
As compressed data is binary, a BytesIO object needs to be used to
handle the data. It was in I79703aa4d3a3d1df5cea546cb18152305c3b0cd4
replaced with io.StringIO which only works on strings (or unicode in
Python 2). Before dd558f287f6e786a41a881b0956e43130260a385 it was using
StringIO.StringIO which works on bytes and strings but with that change
io.StringIO was using in Python 3.

Bug: T86462
Change-Id: Ib3ff54242b25578a7387b97e4c6a10b6d21268ea