Introduce set_redirect_target method in Page class
It set target of redirect page any given page or page and section.
It also can create the redirect or change an ordinary page to a redirect.

The main reason to do this is that It can be altered to support
set redirect in subclasses like ItemPage or CategoryPage and be used generally.

Added one example in the capitalize_redirects script.

Change-Id: Idbe5a1c3e6bb43afe019823fd5eb120c00d9dbec