Do not change pipe links with r# choice
When there is a pipe link, r# as well as # choice options
shouldn't change it. # should always create a pipe link if there
isn't one (which is the given behaviour). But #r should not replace
an already given pipe link with the new target link only but keep
the link text.

Change-Id: I54cc7beec8d9a2037b28d4763dd57a44823fa457