[FEAT] Category: Get newest pages
This returns the pages which have been added to a category ordered by
the creation date from newest to oldest. To do that it needs to cache
all pages and check on each iteration if any previously cached page was
created after the current page is added. If that is the case the cached
pages are newer than the current page and any other page which will be
checked as the generator progresses as they are ordered by addition to
the category and only pages are returned which were added to the
category before the current page was added.

Change-Id: I9bb3f74bbe2e3319ed2dbcdefab414a3204c2c32