[FIX] Upload: Fixing uploads
When there is a page name problem, it reports it and it might then exit
the loop. But if there is no problem it doesn't continue but asks the
user and the loop continues.

It also calls upload_file internally and thus preserving forwarding the
URL, as upload_image was assuming all URLs are a string.

Change-Id: I0020e214db1cf0e276d9267c94d49859802b0a6e