[IMPROV] Deprecate nocapitalize and case
Instead of a static list in the Family file the API should be queried.
All scripts use the Namespace object and APISite itself does check the
Namespace object of the main namespace. The Family class does use a
property which buffers the current state but reads from it will issue a
deprecation warning.

This also deprecates case() as the namespace case should be used if
possible (which is now always usable).

Change-Id: If33ee82d5d9bf657ee36139c409f347b622d3b71