RCStream: Ignore changes after total is reached
There could be 'on_change' messages queued that are still processed
before the thread fully shuts down. This prevents those messages from
appearing in the queue.

As far as I know, the message pump is single threaded, so the previous
on_change call is guaranteed to finish before the next one is called.

Bug: T85716
Change-Id: Ibe6aba6a42028568dbd956da14168fec1bbd920e