Replace getImagePageHtml() with fileIsShared()
ImagePage.getImagePageHtml() was used in the original 2007 checkimages
script to detect if a local page exists for the same name as a file on
the shared repository.  ImagePage now has a cheaper method fileIsShared()
which achieves the same goal.

Likewise getImagePageHtml() was used in the original 2007 unusedfiles
script to detect if the rendered page contained:
   <table id="mw_metadata" class="mw_metadata">
which indicated that it had an image linked to it.  Using fileUrl()
and fileIsShared() achieves the same goal.

Change-Id: I5fd3f34b9081fdbb20b853052fe1c4c30e360d93