[FIX] replace: Allow tuples of two again
With babc7cef9b7805b0bb8c2b5abab4b715e7fc1e7a the replacements needed to
be a list of Replacement instances, but some scripts might call it still
with a list of tuples. This now supports any sequence as long as it has
two entries in it and converts them into Replacement instances.

Bug: T90016
Change-Id: I257badc80a70fc836b9d452402bae75faf2141b1