A BaseSite subclass for non MW sites.
Currently the class only contains a constructor and __getattribute__.
__getattribute__ method returns an attribute if present. This is done
using a whitelist and not hasattr as hasattr uses getattribute of the
super which goes into an infinite recursion. No attributes have been
implemented yet and NotImplementedError is raised if the attribute is
not present in the whitelist. A corresponsding unittest has been
written for this class.

Change-Id: I5d182c6dc4b96b6f06b87ef320bdaec44c8945b8