[IMPROV] login: Simplify reading password
Instead of checking each variant of the tuple a default tuple can be
easily created by inverting the contents and appending the family name
and code if necessary. This makes the condition checking whether the
line applies easier as it doesn't need to care about whether code and
family name are set.

It now also explains which line has an invalid format and explicity
states that the file must be encoded in either UTF-8 or ASCII.

Change-Id: Iff025c267a64b5574c289a3a9ec1350e65728e4d