ParamInfo support for 1.11-1.14
Paraminfo was introduced in 1.12, and until 1.15 it did not include
the 'main' module, which is used to obtain the list of action modules,
or the 'pageset' module.

To workaround this, the API help is parsed to obtain basic parameter
information that pywikibot depends on.

Also paraminfo pre 1.14 does not include limit attributes on parameters
which can have multiple values.

Change-Id: I8a7ae7147554e87b9b097684d4381988ea541b03