[IMPROV] api: Use query module explicitly
When there is a query module and action module with the same name the
ParamInfo module is with a4951559 able to differentiate between both.
With this change ParamInfo chooses the action module if the query module
wasn't chosen explicitly using the module path instead of module name.

To avoid future problems with ambiguous names this patch is explicitly
using the full module path. So this will make it easier to deprecate the
implicit query module selection.

Change-Id: I9e7f8d15d2b469c312c599437b1d44509e9281a7