[FIX] CosmeticChanges: Use correct parameters
When in bce9ad9c6624b6ec914a2607873a9f845700db4a the CurrentPageBot was
introduced CosmeticChanges was also converted and using the new
put_current method but that did change the newtext parameter to new_text
which wasn't reflected in the script. It also accidentally activated
that CosmeticChanges shows the diff too.

Change-Id: I74a3ec50bcd2a0d209b94aa435f67d1ef30da506