[FIX] Return exception instead of raising it
Instead of raising exceptions it should be returned so the data is set
to that exception. It also tests for that case.

It explicitly requires data after the request. One assertion is in the
thread after the lock is released and one assertion is in fetch and
happens also after the lock is released.

Bug: T94993
Change-Id: I955f1c91ae2dc31e404f05ee776d4184882c76f2