Adding commands.log generating function in core.
The compat version contains history of all command which were used for
running bot. It is useful when we want to run some more complicated
command again.There is no such log in core branch. As such I have added
this enhancement as part of Pywikibot:Compat to Core Migration.

Bug:T69724
Change-Id: I8e18c07f1b06569ff29cd8cab3900d515c187f1b