[FIX] Request: Log entire error response
The error code and info has been removed from the response so that in
the APIError it does not appear as a keyword argument. But it also logs
the response without both values. While they are existing in the
APIError instance it is confusing that the response looks very empty:

  {u'servedby': u'mw1122',
   u'error': {'help': u'See https://ru.wikipedia.org/w/api.php for '
                       'API usage'}}

This patch does not remove both keys from the dictionary but skips both
when the keyword arguments are generated.

Change-Id: I9ab89bda4b90e1c14fe25b9fa728f69d2da5b190