[FEAT] Request: Recover from bad tokens
When a badtoken error happens it checks all parameters if one of them is
using a cached token value and then buffers the type for that the token
value is used. It then rerequests all tokens for the buffered types and
is then changing those parameters to have the new value.

Bug: T61678
Change-Id: I0c2f7b3e8726e26c68651bd5714f6f42d1f3bedf