[TOUCHUP] Properly run the no user test
Originally the username in the test was a str but with 1e54a7d6 this
changed into a unicode and added the u-prefix to the output. That patch
fixed the problem by casting the name into str prior to adding it. But
that defies the usage of %r so this patch tries to solve the problem
properly by not casting ALL names into str but only the tested name.

Using str instead of unicode actually uncovered a bug which is already
fixed by a0a194d5.

Change-Id: I3526cb3b29c76c18b92f80ba810b427d8d757617