[IMPROV] Remove usage of 'case' method
In acf2280a the 'case' method of APISite has been deprecated in favor of
the namespace's case value or using the siteinfo directly. This is
removing the usage from pagegenerators_tests and only validating the
case method returns the siteinfo value in site_tests.

It modifies the pagegenerator test testing the TextfilePageGenerator so
that the data is now first orderd by how they appear in the file and
then by the case sensitivity.

Change-Id: I3ebdfa6ffdbfae8e0fdaa533f7283e8a2acab707