[FIX] config2: Support unicode args
With 1e54a7d6 strings like '-dir' are unicode and whenever two strings
are compared and one of them is unicode one of them gets converted. In
this case the argument got converted into unicode using the ASCII codec
which fails if there are non-ASCII entries. This forces '-dir' back into
unicode.

Bug: T95803
Change-Id: Ib899ca2d8a84e5580f095046e91cdd5745da76c0