[FIX] reflinks: Expect that urlopen return bytes
With Python 3 an error occured on using the regex because it only allows
that either all three are bytes or are str but the result from urlopen
is bytes.

On Python 2 all strs are actually bytes by default so it worked fine
there. But with unicode_literals in 1e54a7d6 the replacement and pattern
changed into unicode. While the pattern works in Python 2 regardless the
replacement will force that the searched content is converted into
unicode. As that happens implicitly it'll use ASCII and cause an error
when it contains non-ASCII characters.

The getheader method has been removed Python 3 so it is using the
replacement in Python 3.

Bug: T94688
Change-Id: Icf7b5a02d12cea3ec969fc6a3420ea870ea37f0f