[BREAKING] Change default behavior of set_redirect_target
Now there is another argument named 'save' and default is true.

If it's false, it doesn't save the page but when it's true, it saves page
immediately.

Any usage of this method which doesn't determine any additional
argument would change their behavior and saves immediately instead of
just changing the page.text.

It is useful since most of the time, we don't need more changes after
creating or changing a redirect page.

Change-Id: Iaadfa03d6ef7a8252ff4df7fc5281e222642ebc3