[FIX] Expect that the template is a Page instance
The Page method templatesWithParams returns the first value as a Page
instance instead of a string. It's not possible to use equality in such
a case:

  page.templatesWithParams[0][0] == 'foobar'

This will always be False as the value on the left is a Page instance.
Since 55d6232e the method returns a Page instance so it's not possible
to return string instances like compat does without breaking scripts
which already expecte Page instances.

Bug: T97248
Change-Id: I9d4feffc7d081cbcc349111b3e28da2de300649c