Introduce ProofreadPage(Page)
Class ProofreadPage(Page) is introduced in dedicated module
proofreadpage.py.

Page is divided in header, body and footer according to the Proofread
Extension format.

Main application is for wikisource bots.

Purpose of this commit is to introduce the class to start applying it
and collect feedbacks.

The class can be further extended and improved in subsequent patches,
once its usage will give first feedbacks.

Tests have been started as well.

Change-Id: I34c0170133c7896daa2f1f998758c29c1e7748cd