[FEAT] http: Get revision/version if necessary
This changes pywikibot.comms.http to always use the version number even
if it's not cached. The pywikibot.version module is now always using a
custom user-agent to avoid that {version} or {revision} appears in it
and creating an infinite recursion.

To avoid defining the version and revision if they are not appearing in
the user-agent it is using a custom formatter which handles those two
special so that they don't have to appear in format's args and kwargs.
Instead they will be only defined if the formatter finds those in the
string.

Change-Id: Ia7a31cfba4d467118713fbb0122dad932b3a5c39