[FIX] TerminalHandler: Expect str in Python 2
The args of a record are str in Python 2 (so basically bytes). It now
compared the str with unicode which works in Python 3 (as both are the
same there) or Python 2 if the warning message contains only ASCII
characters.

With 1e54a7d6 unicode_literals was added to the module which changed the
strings into unicode in Python 2. And Python converts the string into
unicode when it tries to do a __contains__ test. It decodes it using
ASCII which fails when there are non-ASCII characters in the string.

Bug: T98955
Change-Id: Ia421c085897e14e01a59e3c2a14822b2a7cd0ddd