[FIX] replaceExcept: Replace references iteratively
Instead of searching for back references (\0 or \g<foo>) on the entire
resulting string it just searches over the original replacement string. So when
back references appear in the resulting string itself, they won't be found.

It also allows group numbers in \g back references to act the same as re.sub.

Bug: T99032
Change-Id: Ia4419804b83e67baada277e1493b58e4d2f73bf5