[FIX] replaceExcept: Handle empty matches
When the regex can match the empty string, it will get into an infinite loop as
soon as it reached the end of the string, as it will match the last 0
characters over and over.

This uses the same functionality from re.sub() that it only replaces if it
isn't adjacent to a previous match.

Change-Id: I479d6d5f9ac756b84d93bfd1fd81a2948677787f