Allow relative links to be handled in Link constructor
This patch allows a BasePage object to be passed in lieu of a Site
to the Link constructor. In this case, the link text will be handled
relative to the passed Page. This will allow for subpage links to be
parsed when relative links are used on the page. A new test, test_relative,
has been added to tests/link_tests.py to test this change.

Bug: T57113
Change-Id: Idd6366e82216c130b5143c9356245ff840af41ed