[FIX] i18n: twtranslate returns used code in list
It fixes an issue with 5767d73d, that twntranslate didn't get the language code
which was actually used. With fallback=True twtranslate might use another
language than the language given in the first place. To inform twntranslate
about that, twntranslate uses a list which is modified by twtranslate.

But with 5767d73d it always added the same value again so that the list wasn't
chaned at all.

Change-Id: I7228b1ab19510769c876e8943e92dc666c6b2ded