[FIX] tests: Support Python 3 in metaclass tests
Some tests (date, l10n and logentry) were just using __metaclass__ which is not
supported by Python 3. The metaclasses in aspects and script_tests were using
six on Python 3 to avoid that. This adds a generic decorator to tests.utils
which looks for __metaclass__ in the class itself and then uses six's
add_metaclass (like in aspects and script_tests) to add this metaclass.

Change-Id: I8cb46bffb39ef4640ec523e596fba5fbf48fbd8d