[FEAT] Request: Separate MW API and PWB API params
Instead of using kwargs, which could clash with already existing normal
parameters it uses a separate argument which takes all the parameters used for
the API.

To support the mixed mode in other classes forwarding kwargs directly to
Request they can use 'clean_kwargs' which separates the API parameters and
Python arguments to use the new style. Subsequent changes must reflect that
then. This will not break backwards compatibility, just that implementations
not using 'clean_kwargs' will only be able to use one mode.

It also adds support for Page instances in parameters and adds a function in
tools to merge multiple dicts if they have unique keys.

The APISite instance also gets '_request', '_simple_request' and
'_request_class' methods in case a script want to use their subclass of
Request. The QueryGenerator also uses '_request' and '_request_class'. But for
backwards compatibility it allows both modes (so it's converting kwargs mode to
parameters mode). With that QueryGenerator uses CachedRequest when an expiry
parameter is set which breaks any usage with such an API parameter unless they
use the parameters mode.

Bug: T70988
Change-Id: Ic259420c312d4efe4f14c772b369f86e6ed2e819