Add 'need_extension' decorator
This patch adds a decorator to ensure that a particular extension
required for an API call on a site is available on that site.
Two methods requiring the Echo extension were converted to use this
decorator instead of the existing if test, and a ProofreadPage-related
method now uses the decorator instead of raising an error.

Change-Id: I0e609face41ff2b9d4a1169f736f04966c1b71c1