Split script tests into help and simulate tests
List 'version' as an unrunable script, as it does
not support global options.

Four auto-run scripts are moved from expected failures
to be allowed failures, as their output is highly
variable depending on existence of site templates or
categories, and may not emit output within the time
given on wikis with large backlogs.

Change-Id: I7f29107f2d4db5e6c639f349e83ce4705d478034