[FIX] open_compressed: Wrap BZ2File in Py 2.7
Although the documentation specifies that BZ2File is compatible with the
with-statement since Python 2.7 and the code for both gzip and BZ2File were
changed in the same commit in cpython, only since Python 2.7.1 this is
supported.

It also separates the tests so it's easier to see if only one specific
compressor has a problem. It also tests that the wrapper does not block any
exception.

The docstring is also improved to reflect the use_extension parameter.

Bug: T101649
Change-Id: I3f7024746164cc1bb1d89f2ac52dfa41ca08a3a6