Test extras on Appveyor builds
Skip installation of oursql, as it implicitly depends on Cython
and fails if it is not present, there are not any tests
using the mysql datasource, and the main usage is on the UNIX
Wikimedia labs where the databases are locally available.

Also Cython is large and requires a compiler, which many
Windows users do not have installed.

Change-Id: Ieef7b7480297fc82ddcfebfd12251006fc99efe2