[FIX] upload: Use correct offset
When a warning occurs and there is no offset in the request it either means
that this is not a chunked upload or that it's the first chunk. If it's not a
chunked upload it should select the offset 'False' and otherwise '0'. The patch
introducing finishing/continuing uploads (d0d1c830) always uses '0' which means
that after a warning with an unchunked upload it'll use a chunked upload (which
does not work for URL uploads).

Using an offset of '0' is in theory wasting the first chunk which may have been
uploaded successfully but unfortunately it might be that not the entire chunk
was transmitted. (There are cases where the returned offset did not match with
the old offset + chunk size)

Change-Id: Ic1e5cb8fee9d4d8ead591947f36c8436540e1b73