[BREAK] [FEAT] SingleSiteBot and MultipleSitesBot
Several bots can only work with pages on one site as they initialize stuff
which is site specific. In theory page generators can return pages for
different sites which can cause errors.

This also adds MultipleSitesBot which is specifically for bots that can work on
multiple sites in one run and it's not allowed to cache the site in self.site
there.

This converts all scripts which utilize Bot's run method where both new
subclasses will have an effect.

It breaks the usage of CurrentPageBot which now uses the BaseBot as the
superclass which does no site management. Instead it should be used with one of
the new subclasses or the site management should be done by the script itself.

Change-Id: I5e8759ccaef0063d6d6f9842bf3d3af3baa0f567