Extract sys.version_info tuple once
It is more efficient to check a bool for Python 2 vs 3.
The tools variables PYTHON_VERSION and PY2 allow for consistent
usage.  They have not been utilised in this changeset where
sys.version_info is used only during startup and only a few times
during the startup, as there is no significant performance benefit.

Also changed str to its equivalent bytes on Python 2.

Change-Id: I0a29bed4b6239b55f5391ead07ea2fffcaad4855