[FEAT] config: support different types
After reading the user-config it compared the types and issued warnings when
they changed (with a few exceptions). But some values like 'socket_timeout'
support multiple types and it warns if it changed the type to a valid one but
not the one used by default.

This adds the possibility to define a method which takes the original value and
returns a parsed value and issues an exception otherwise.

It also reduced the number of exceptions by default. Previously configuration
values could be changed from int to float or bool without warning and it would
allow int values instead of float or bool values. Of those only allowing int
values instead of a float is sensible. But allowing floats instead of ints
might result in problems. And if it was a boolean it is not sensible that they
just may be replaced by int.

If variables still need that they can just implement their own handler.

Bug: T103154
Change-Id: I435274187c0635ecb5bf2bb73ff9cb2de898d19e