[FEAT] api tests: Test APIMWException
The patch adds tests for the APIMWException. To reliably cause such an
exception it patches the http import from the pywikibot.data.api module to its
own class which returns predefined data.

Because there might be some setting up involved before the actual request is
done that should be tested, the tests using a Page instance do a dummy call
before.

Change-Id: I6165a073d74916c9dc42f7667ae18bdd027b1176