[FEAT] input_choice: modular options
This makes input_choice more modular an thus allowing nested options or
numbers as options. It removes the check that each option is unique so the
caller needs to take care of that. The options are now mutable so that they can
represent changes. For example when the list changes and it's possible to
select one item from the list.

Change-Id: Ib922bcdb77ee34e8d0268ded2b003e7e6a9005b4