[BREAK][FIX] Redesign of InteractiveReplace
It adds the Choice class which can be (along with ChoiceExceptions) in a
separate list to have their own handling and allow to add the same kind of
choice multiple times.

The 'y' shortcut has been removed as it is odd because the user doesn't get the
same shortcut presented for all the choices. It also doesn't simplify anything
as there is also a default option (which might be different). Apart from that
it was buggy and it assigned always the last allowed replace option instead of
the first which meant that the replace option using 'y' wasn't actually
available (as it has been mapped the last one which also has the ordinary
shortcut).

Change-Id: I56b32a0f69147fb46772b8a1dffdaef5b19931de