[IMPROV] APISite: Deprecate language method
The APISite.language() method and APISite.lang property return the exact same
thing so it does make sense to remove one of them. As BaseSite itself has a
lang property already it makes sense to deprecate the language() method.

This makes the lang property in APISite more visible and it's clearer that
APISite actually overrides the implementation from BaseSite. Searching for
'def lang(' for example would only yield BaseSite's implementation.

Also fixed several usages to use the code instead of the language.

Change-Id: I938c6b587cc020ba401fb09883520bbb2a5367c0