[FEAT] NamespacesDict to replace Namespace methods
Instead of having static/classmethods in Namespace use a dict which stores the
Namespace instances and allows general handling of them. This reuses the
_NamespacesDict introduced in ae69d8dc. This makes doing lookups easier as it's
not necessary to import the Namespace class and having to manually define on
which namespaces it works.

This makes also ns_index unnecessary as it's relatively easy possible to
actually get the Namespace using the name and a site instance now.

It also adds a new TestCase class which patches assert calls to easily check if
the deprecation message was issued after each assertion.

Change-Id: I74507af6f22cf09fa161f5df07e1dd90cb399759