[BREAK][FEAT] Support strings in replace_links
It breaks parts of replace_links that when it the replacement may be a string
and it will be replaced directly and not converted into a link. This only
happens when the replacement was a callable and not a sequence so that it's
still possible to quickly replace links by using a sequence. It won't support
bytes to avoid encoding issues.

On start of replace_links it'll convert the string into a link if it's using a
sequence and not a callable.

This change is necessary for selflink which also supports replacing a link with
any text (in that case the bold page title). It now subclasses the unlink script
as almost everything can be reused.

Change-Id: I54c4015795d8190606454200be23c07016459d26