[FIX] Replace with Link if all or only target selected
The options 'Change link target' and 'Change complete link' previously returned
just a string as a shortcut. But with ad2cf8aa if a callback returns a string
it's inserted directly.

It adds the option to replace the target and the section but not the label and
to replace everything to disambredir.

It also changes the default title behaviour in replace_link that it's using
also the section like a link in MW would behave. If the replacement is
Link('Foo#Bar') it's interpreted as [[Foo#Bar]] so that the title is Foo#Bar.

Bug: T104805
Change-Id: Ic6f3d1dd0d2e56a639ebf74e89834cf02d7b224a