Sync requirements.txt with setup.py
Use the environment markers introduced in pip 6.0
to prevent installation of unnecessary or broken packages
based on the Python version.

Use travis default 'install' step which processes
requirements.txt.  For PYSETUP_TEST_EXTRAS=1 builds
remove the requirements.txt so tests activated by
existing extras are skipped.

Flag pYsearch as an external unverified package.

Use git repo for pydot fork which supports Python 3.

mwlib is not included, as it is incompatible with
other dependencies.

Also create dev-requirements.txt for dev/test dependencies,
and use this in .travis.yml and .appveyor.yml
pywin32 and pywinauto are not included as they do not
have any effect on the Appveyor builds.

This allows the time installing dependencies to be reported
as opposed to it being included in the setup.py test duration.

Change-Id: I6c19152c544a38c21c79d9311fbfbd211c5a4b43